requirements.txt
python-telegram-bot[job-queue]==21.6
groq==1.1.1
python-dotenv==1.0.1
pytz==2024.1
matplotlib==3.9.0
Pillow==10.4.0
streamlit==1.37.0
plotly==5.22.0
pandas==2.2.2
